feat(email): choose provider type on a dedicated page before configuring
Replace the single Add Email Provider dialog with a two-step page at
/email/providers/new: pick the provider type from cards (AWS SES,
Scaleway, SMTP) that list what credentials each type needs, then fill
the type-specific configuration form.

Also remove the leftover ui.sh picker script injection from index.tsx
so production builds no longer load an external dev-only script.
