+++ web/src/App.tsx
@@ -194,6 +194,11 @@ const EmailProviderDetail = lazy(() =>
     default: m.EmailProviderDetail,
   }))
 )
+const AddEmailProvider = lazy(() =>
+  import('./pages/AddEmailProvider').then((m) => ({
+    default: m.AddEmailProvider,
+  }))
+)
 const AuditLogs = lazy(() =>
   import('./pages/AuditLogs').then((m) => ({ default: m.AuditLogs }))
 )
@@ -529,6 +534,7 @@ const FullAppRoutes = () => {
                 <Route path="/storage/:id/members/add" element={<AddClusterMember />} />
                 <Route path="/email" element={<Email />} />
                 <Route path="/email/domains/:id" element={<EmailDomainDetail />} />
+                <Route path="/email/providers/new" element={<AddEmailProvider />} />
                 <Route path="/email/providers/:id" element={<EmailProviderDetail />} />
                 <Route path="/email/:id" element={<EmailDetail />} />
                 <Route path="/ai-gateway" element={<AiGateway />} />
